requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
langchain>=0.1.0
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
python-docx>=0.8.11
pymupdf>=1.23.0
pillow>=10.0.0
pytesseract>=0.3.10
typer>=0.9.0
rich>=13.6.0
pytest>=7.4.0

